Remove OpenGL hook reheck in Windows overlay.
Originally added by slicer in c865f2f3.

MinHook cannot currently do this. Since OpenGL
is the only user of the HardHook::check() functionality,
I think it's reasonable to replace this with a call to
ods().

If we run into an app where there is a problem, we can
work something into MinHook.