Add CONFIG(tlsv1+) option to add support for TLS 1.1 and TLS 1.2 (requires patched Qt 4)
With this new config option, Mumble and Murmur now speak
TLS 1.0, TLS 1.1 and TLS 1.2 (and greater, depending on
your OpenSSL version).

As documented in the INSTALL file, this new config option
requires a patched Qt.

This means it's mostly useful for our own binary packages, but
that's better than nothing.