Log overlay helper process errors in OverlayPrivateWin.
OverlayPrivateWin used to only listen for process
termination. Technically, this should work just fine.

This change makes OverlayPrivateWin listen for process
errors as well, by subscribing to the error() signal
of the QProcess of each of our helpers.

The idea is that this could provide us helpful error
messages when troubleshooting issues where overlay
helpers are exiting untimely.