SSL: implement OpenSSL cipher list format reader.
This commit implements MumbleSSL::ciphersFromOpenSSLCipherString(),
which reads a line in OpenSSL's cipher list format and returns a list
of QSslCiphers for use with QSsl.

This function will be used for implementing user-configurable cipher
suites in Mumble and Murmur.

This commit also changes our precompiled headers slightly by rearranging
the OpenSSL includes. This is done because the ssl.h header on Windows
requires winsock2.h (and perhaps others). By moving the includes, we
ensure that winsock2.h is included.