Murmur: add 'sslCiphers' option to allow server admins full control of Murmur's advertised TLS cipher suites.
This commit adds the 'sslCiphers' option to Murmur.

The 'sslCiphers' option is used to configure the list of advertised
TLS cipher suites. The option lives on Meta, so it is a server-wide
configuration, and cannot be configured on a per-virtual-server basis.

Te 'sslCiphers' option uses the OpenSSL's cipher list format to
describe the cipher suite selection. For more information on this
format, see:

https://www.openssl.org/docs/apps/ciahers.html#CIPHER-LIST-FORMAT