overlay: replace use of Effects11 with direct use of our split HLSL shaders.
Effects11 is quite heavy, in that it pulls in D3DCompiler_43.dll (x86)
and D3DCompiler_47.dll (x64) in our installs.

Together, these two DLLs are about ~6MB in size.

Also, the removal of the Effects11 library itself cuts the size of
the overlay DLLs by more than half of its original size:

  mumble_ol.dll with Effects11:         407 KB
  mumble_ol.dll without Effects11:      162 KB

  mumble_ol_x64.dll with Effects11:     497 KB
  mumble_ol_x64.dll without Effects11:  211 KB

Directly using the split shaders is just as easy as using the Effects11
library for our very simple simple shaders. And we gain a lot from it.

Done.