overlay: replace use of Effects11 with direct use of our D3D11 overlay vertex and pixel shader.
Effects11 is quite heavy, in that it pulls in D3DCompiler_43.dll (x86)
and D3DCompiler_47.dll (x64) in our installs.

Together, these two DLLs are over 6MB in size.

Directly using the vertex and pixel shaders is just as easy
as using the Effects11 library for our simple shader, and
we gain a lot from it.

Done.